Fixed mitochonidrial ids, parents and circular references.
A number of problems were fixed:

SNOR_00002 had itself as a parent (fixed to SNOG_00002)
There were a number of mitochondrial orfs that were not assigned locus numbers. The script that originally produced the gff had simply created the mRNA feature and given the parent as the last gene it knew about. This caused atp6, for example to have two extra mRNA features that lay outside the gene region. The five mRNA features that had incorrect parents were corrected by adding the new genes mito_orf_G_01 to mito_orf_G_05.
Also, a number of mRNA IDs were exactly the same as their parent's (the gene name). This was corrected by adding ".t0d" to the end of the mRNA ID, where d is a digit.